requirements.txt
flask
flask-cors
openai
python-dotenv
pdfplumber
pytesseract
pdf2image
gunicorn